backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
openai>=1.0.0
python-dotenv==1.0.0
numpy==1.24.3
pandas==2.0.3
pytest==7.4.3  # For running tests 